requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pydantic
pydantic-settings
supabase
python-dotenv
openai
pypdf2
python-docx
aiofiles
httpx
